+++ conversation/handler.go
@@ -1,7 +1,7 @@
 package conversation // import "heytobi.dev/fuse/conversation"
 
 import (
-	"github.com/sirupsen/logrus"
+	"context"
 	"heytobi.dev/fuse/telegram"
 )
 
@@ -19,7 +19,7 @@ type bot interface {
 type Handler struct {
 	bot             bot
 	activeSequences map[int64]Sequence
-	defaultResponse string
+	defaultSequence Sequence
 }
 
 // NewHandler ...
@@ -31,32 +31,21 @@ func NewHandler(bot bot) *Handler {
 }
 
 // Handle handles every incoming message that doesn't have a dedicated handler.
-func (h *Handler) Handle(update *telegram.Update) error {
+func (h *Handler) Handle(ctx context.Context, update *telegram.Update) error {
 	if update != nil && update.Message != nil {
 		// check if there is an active sequence for this user, delegate to that sequence if there is one.
 		if sequence, hasActiveSequence := h.activeSequences[update.Message.Chat.ID]; hasActiveSequence {
-			err := sequence.Process(update)
+			err := sequence.Process(ctx, update)
 			if err != nil {
 				return err
 			}
 			return nil
 		}
 
-		if h.defaultResponse != "" {
-			result, err := h.bot.SendMessage(&telegram.SendMessageRequest{
-				ChatID: update.Message.Chat.ID,
-				Text:   h.defaultResponse,
-			})
-
+		if h.defaultSequence != nil {
+			err := h.defaultSequence.Process(ctx, update)
 			if err != nil {
-				logrus.WithError(err).Error("failed to send telegram message")
 				return err
-			}
-
-			if !result.Successful {
-				logrus.WithFields(logrus.Fields{
-					"response": result,
-				}).Warn("send message result was false")
 			}
 		}
 	}
@@ -64,9 +53,9 @@ func (h *Handler) Handle(update *telegram.Update) error {
 	return nil
 }
 
-// RegisterSequence registers the active sequence for the given user. New registrations always override any already
-// registered sequence, and there can only be 1 or no active sequences for each user, tracked by the telegram chat ID.
-func (h *Handler) RegisterSequence(chatID int64, sequence Sequence) error {
+// RegisterActiveSequence registers the active sequence for the given user. New registrations always override any already
+// registered sequence. There can be at most 1 active sequences for a user, tracked by the telegram chat ID.
+func (h *Handler) RegisterActiveSequence(chatID int64, sequence Sequence) error {
 	h.activeSequences[chatID] = sequence
 	return nil
 }
@@ -78,8 +67,8 @@ func (h *Handler) DeregisterActiveSequence(chatID int64) error {
 	return nil
 }
 
-// WithDefaultResponse sets a default response for cases where there is no active sequence.
-func (h *Handler) WithDefaultResponse(response string) *Handler {
-	h.defaultResponse = response
+// WithDefaultSequence sets a fallback sequence for messages without a dedicated sequence.
+func (h *Handler) WithDefaultSequence(sequence Sequence) *Handler {
+	h.defaultSequence = sequence
 	return h
 }